refactor(tests): reorganize test structure for clarity and scalability
- Grouped tests by type (unit, integration, e2e) and module (sdk, server).
- Renamed directories for better readability and alignment with best practices.
- Removed unnecessary __init__.py files from test directories.
- Centralized shared test configurations in conftest.py for reusability.
- Improved flexibility for running module-specific and type-specific tests.

This refactor enhances maintainability, simplifies navigation, and prepares the test suite for future expansion.
