+++ src/common/exceptions.py
@@ -1,4 +1,4 @@
-from ..utils.logger import logger
+from .logger import logger
 
 
 class ApiError(Exception):
