+++ src/sdk/features/messages.py
@@ -1,10 +1,10 @@
 from typing import Dict, List
 from ..client import ApiClient
 from ..schemas.messages import CreateMessageRequest, Message, ListMessagesResponse
-from ..utils.validators import validate_request, validate_response
-from ..utils.exceptions import handle_exceptions
-from ..utils.logger import logger
-from ..utils.validators import verify_signature
+from src.common.validators import validate_request, validate_response
+from src.common.exceptions import handle_exceptions
+from src.common.logger import logger
+from src.common.validators import verify_signature
 
 
 class Messages:
