+++ tests/integration/server/test_webhook.py
@@ -2,7 +2,7 @@ import json
 from fastapi.testclient import TestClient
 from config import settings
 from src.server.app import app
-from src.sdk.utils.validators import generate_signature
+from src.common.validators import generate_signature
 
 client = TestClient(app)
