+++ tests/unit/sdk/test_client.py
@@ -1,7 +1,7 @@
 import pytest
 from unittest.mock import patch, MagicMock
 from src.sdk.client import ApiClient
-from src.sdk.utils.exceptions import UnauthorizedError, NotFoundError, ServerError, ApiError
+from src.common.exceptions import UnauthorizedError, NotFoundError, ServerError, ApiError
 from config import settings
